调整回tsc打包
