+++ rollup-plugin/single-file/watchRollup.js
